+++ internal/ingress/env.go
@@ -1,0 +1,5 @@
+package ingress
+
+import "os"
+
+func stdGetenv(k string) string { return os.Getenv(k) }